requirements.txt
strands-agents[bidi]
strands-agents-tools
fastapi
uvicorn[standard]
websockets
requests
